+++ src/Executor/ClickHouse/ClickHouseExecutor.php
@@ -1,0 +1,36 @@
+<?php
+
+/*
+ * This file is part of the FiveLab Ruler package
+ *
+ * (c) FiveLab
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code
+ */
+
+declare(strict_types = 1);
+
+namespace FiveLab\Component\Ruler\Executor\ClickHouse;
+
+use FiveLab\Component\Ruler\Executor\ExecutorInterface;
+use FiveLab\Component\Ruler\Node\Node;
+use FiveLab\Component\Ruler\Operator\Operators;
+use FiveLab\Component\Ruler\Query\ClickHouseQuery;
+
+/**
+ * @implements ExecutorInterface<ClickHouseQuery>
+ */
+readonly class ClickHouseExecutor implements ExecutorInterface
+{
+    public function __construct(private ClickHouseVisitor $visitor, private Operators $operators)
+    {
+    }
+
+    public function execute(object $target, Node $node, array $parameters): void
+    {
+        $where = $this->visitor->visit($target, $node, $parameters, $this->operators);
+
+        $target->andWhere($where, $parameters);
+    }
+}