+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+v1.4.0
+------
+
+* Added `FiveLab\Component\Ruler\Target\ClickHouseTarget` for build the conditions for ClickHouse.
+* Added `FiveLab\Component\Ruler\Query\ClickHouseQuery` for use it as target for ClickHouse.
+
+v1.3.1
+------
+
+* Add support PHP `8.4`.
+* Fix implicit nullable parameters in `FiveLab\Component\Ruler\Ruler`.
+
 v1.3.0
 ------